Account for no alt
